# [routing] Routing performance optimization. Stop checking twin features geometry on the same mwm versions.