@bykoianko 2020-10-22
> теоретически с самосборными мвм можно получить проблемы или если два раза пересоберём с одной датой (в этом году такое было), но в продакшен это уходить не должно и код ради таких случаев усложнять наверное не стоит

Думаю, да. Вообще, это случай отличный от обычного workflow. И если он случится, то это крайне редкие креши. Т.е. мы должны отклониться от типичного workflow + должно оч сильно не повезти.